# config/swagger.nodos.json
{
    "openapi": "3.0.0",
    "info": {
       "version": "1.0",
       "title": "API red de sensores",
       "description": "Documentacion de los endpoints disponibles para el proyecto de red de sensores",
       "license": {
          "name": "ITESO",
          "url": "https://www.iteso.mx/"
       }
    },
    "paths": {
...
       "/lecturas": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Inserta una lectura en la base de datos",
             "operationId": "postLectura",
             "parameters": [
                {
                   "name": "cmd",
                   "in": "query",
                   "description": "Lectura de nodo a insertar",
                   "required": true,
                   "type": "string"
                }
             ],
             "responses": {
                "201": {
                   "description": "Lectura creada satisfactoriamente"
                }
             }
          }
       },
...
       "/lecturas/id/{id}": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Obtiene una lectura con un ID específico",
             "operationId": "getLecturaId",
             "parameters": [
                {
                   "name": "id",
                   "in": "path",
                   "description": "Id a consultar la lectura",
                   "required": true,
                   "type": "string"
                }
             ],
             "responses": {
                "200": {
                   "description": "Lectura obtenida satisfactoriamente"
                }
             }
          },
          "put": {
             "tags": [
                "Lecturas"
             ],
             "security": [
                "BearerAuth"
             ],
             "description": "Actualiza una lectura con un ID específico",
             "operationId": "putLecturaId",
             "parameters": [
                {
                   "name": "id",
                   "in": "path",
                   "description": "Id a actualizar la lectura",
                   "required": true,
                   "type": "string"
                },
                {
                   "in": "header",
                   "name": "Authorization",
                   "schema": {
                      "type": "string"
                   },
                   "description": "Bearer token",
                   "required": true
                }
             ],
             "responses": {
                "200": {
                   "description": "Lectura actualizada satisfactoriamente"
                }
             }
          },
          "delete": {
             "tags": [
                "Lecturas"
             ],
             "security": [
                "BearerAuth"
             ],
             "description": "Elimina una lectura con un ID específico",
             "operationId": "deleteLecturaId",
             "parameters": [
                {
                   "name": "id",
                   "in": "path",
                   "description": "Id a la lectura",
                   "required": true,
                   "type": "string"
                },
                {
                   "in": "header",
                   "name": "Authorization",
                   "schema": {
                      "type": "string"
                   },
                   "description": "Bearer token",
                   "required": true
                }
             ],
             "responses": {
                "200": {
                   "description": "Lectura eliminada satisfactoriamente"
                }
             }
          }
       },
       "/lecturas/n/{nodo}": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Obtiene las últimas lecturas de un nodo en específico",
             "operationId": "getLecturasNodo",
             "parameters": [
                {
                   "name": "nodo",
                   "in": "path",
                   "description": "Id del nodo a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "count",
                   "in": "query",
                   "description": "Número máximo de lecturas, por defecto 100",
                   "required": false,
                   "type": "integer",
                   "default": "100"
                }
             ],
             "responses": {
                "200": {
                   "description": "Lecturas obtenidas satisfactoriamente",
                   "content": {
                      "application/json": {
                         "type": "array"
                      }
                   }
                }
             }
          }
       },
       "/lecturas/dia/{nodo}/{año}/{mes}/{dia}": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Obtiene las lecturas de un nodo en específico en un día",
             "operationId": "getLecturasNodoDia",
             "parameters": [
                {
                   "name": "nodo",
                   "in": "path",
                   "description": "Id del nodo a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "año",
                   "in": "path",
                   "description": "Año a consultar las lecturas",
                   "required": true,
                   "type": "integer"
                },
                {
                   "name": "mes",
                   "in": "path",
                   "description": "Mes a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "dia",
                   "in": "path",
                   "description": "Día a consultar las lecturas",
                   "required": true,
                   "type": "integer"
                }
             ],
             "responses": {
                "200": {
                   "description": "Lecturas obtenidas satisfactoriamente",
                   "application/json": {
                      "type": "array"
                   }
                }
             }
          }
       },
       "/lecturas/semana/{nodo}/{año}/{mes}/{dia}": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Obtiene las lecturas de un nodo en específico en una semana",
             "operationId": "getLecturasNodoSemana",
             "parameters": [
                {
                   "name": "nodo",
                   "in": "path",
                   "description": "Id del nodo a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "año",
                   "in": "path",
                   "description": "Año a consultar las lecturas",
                   "required": true,
                   "type": "integer"
                },
                {
                   "name": "mes",
                   "in": "path",
                   "description": "Mes a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "dia",
                   "in": "path",
                   "description": "Día a consultar las lecturas",
                   "required": true,
                   "type": "integer"
                }
             ],
             "responses": {
                "200": {
                   "description": "Lecturas obtenidas satisfactoriamente",
                   "application/json": {
                      "type": "array"
                   }
                }
             }
          }
       },
       "/lecturas/mes/{nodo}/{año}/{mes}": {
          "get": {
             "tags": [
                "Lecturas"
             ],
             "description": "Obtiene las lecturas de un nodo en específico en un mes",
             "operationId": "getLecturasNodoMes",
             "parameters": [
                {
                   "name": "nodo",
                   "in": "path",
                   "description": "Id del nodo a consultar las lecturas",
                   "required": true,
                   "type": "string"
                },
                {
                   "name": "año",
                   "in": "path",
                   "description": "Año a consultar las lecturas",
                   "required": true,
                   "type": "integer"
                },
                {
                   "name": "mes",
                   "in": "path",
                   "description": "Mes a consultar las lecturas",
                   "required": true,
                   "type": "string"
                }
             ],
             "responses": {
                "200": {
                   "description": "Lecturas obtenidas satisfactoriamente",
                   "application/json": {
                      "type": "array"
                   }
                }
             }
          }
       },
...
    },
    "servers": [
       {
          "description": "Host",
          "url": "http://papvidadigital-test.com/api"
       }
    ],
    "components": {
       "securitySchemes": {
          "BearerAuth": {
             "type": "http",
             "scheme": "bearer"
          }
       },
...
    }
 }